refactor(providers): centralize encrypted V2 route eligibility in one helper
