+++ src/codex/subagent-model-fallback.ts
@@ -30,16 +30,13 @@ import { isCodexAccountPaused } from "./account-pause";
 import { slugEquals } from "../providers/slug-codec";
 import { isThreadSpawnRequest } from "../server/effort-policy";
 import { PROVIDER_REGISTRY, type InboundWire } from "../providers/registry";
-import { getOAuthCredentialApiBaseUrl } from "../oauth";
 import {
-  canReceiveEncryptedV2AgentTasks,
   CODEX_FORWARD_BASE_URL,
   OPENAI_CODEX_PROVIDER_ID,
   isCanonicalOpenAiForwardProvider,
 } from "../providers/openai-tiers";
 import { routeModel, type RouteResult } from "../router";
-import { resolveFinalWireProtocolOverride } from "../server/adapter-resolve";
-import { resolveProviderTransport } from "../providers/xai-transport";
+import { canRouteEncryptedV2AgentTasks } from "../providers/encrypted-v2-transport";
 import { sweepExpiredOnWrite } from "../lib/state-store-sweeper";
 import { codexAccountNamespaceForModel } from "./account-namespace-match";
 import { ACCOUNT_GATED_NATIVE_OPENAI_MODELS } from "./catalog/native-models";
@@ -74,14 +71,6 @@ type ModelHealth = {
 const modelHealth = new Map<string, ModelHealth>();
 const quotaPrimedAt = new Map<string, number>();
 const knownProviderIdSet = new Set(PROVIDER_REGISTRY.map(entry => entry.id.toLowerCase()));
-
-function sameUpstreamOrigin(left: string, right: string): boolean {
-  try {
-    return new URL(left.trim()).origin === new URL(right.trim()).origin;
-  } catch {
-    return false;
-  }
-}
 
 function tryRouteFallbackModel(config: OcxConfig, model: string): RouteResult | null {
   try {
@@ -375,27 +364,7 @@ export function selectAvailableSubagentModel(
   for (const candidate of chain) {
     if (nativeFallbackOnly) {
       const route = tryRouteFallbackModel(config, candidate);
-      const resolvedProvider = route
-        ? (() => {
-            const approvedBaseUrl = route.provider.baseUrl;
-            const transportProvider = resolveProviderTransport(
-              route.providerName,
-              route.provider,
-              undefined,
-              route.providerName === "github-copilot"
-                ? getOAuthCredentialApiBaseUrl(route.providerName)
-                : undefined,
-            );
-            if (!sameUpstreamOrigin(approvedBaseUrl, transportProvider.baseUrl)) return null;
-            return resolveFinalWireProtocolOverride(
-              route.providerName,
-              route.modelId,
-              transportProvider,
-              inboundWire,
-            );
-          })()
-        : null;
-      if (!resolvedProvider || !canReceiveEncryptedV2AgentTasks(resolvedProvider)) {
+      if (!route || !canRouteEncryptedV2AgentTasks(route, inboundWire, route.provider.baseUrl)) {
         skipped.push(candidate);
         continue;
       }
